% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/JSTOR_MALLET_topicinfo.R
\name{JSTOR_MALLET_topicinfo}
\alias{JSTOR_MALLET_topicinfo}
\title{Get information about a topic generated by MALLET}
\usage{
JSTOR_MALLET_topicinfo(x, n, threshold = 0.3)
}
\arguments{
\item{x}{the object returned by the function JSTOR_unpack.}

\item{n}{the topic number to get information about (starts at 1)}

\item{threshold}{proportion of topic required for a document to be considered exemplary of that topic}
}
\value{
Returns a list that includes top words, alpha value, and top articles for the selected topic
}
\description{
Gets keywords, DOIs, alpha values and exemplary articles for a topic. Originally written by Andrew Goldstone. For use with JSTOR's Data for Research datasets (http://dfr.jstor.org/).
}
\examples{
## info <- JSTOR_MALLET_topicinfo(x = unpack, n = 1)
}

