% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JSTOR_subset1grams.R
\name{JSTOR_subset1grams}
\alias{JSTOR_subset1grams}
\title{Subset 1gram object}
\usage{
JSTOR_subset1grams(unpack1grams, word)
}
\arguments{
\item{unpack1grams}{object returned by the function JSTOR_unpack1grams.}

\item{word}{The word or vector of words to subset the documents by, ie. use only documents containing this word (or words) in the cluster analysis}
}
\description{
Subsets 1 grams
}

