% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/JSTOR_unpack_multiple_archives.R
\name{JSTOR_unpack_multiple_archives}
\alias{JSTOR_unpack_multiple_archives}
\title{Plot the most frequent words by time intervals}
\usage{
JSTOR_unpack_multiple_archives(mydir = getwd())
}
\arguments{
\item{mydir}{path to directory containing multiple zip files dowloaded from dfr.jstor.org (default is the current working directory)}
}
\value{
Returns a list of two items. First is "wordcounts", a Document Term Matrix of 1-grams, and second is 'bibliodata', a data frame of bibliographic information for all articles.
}
\description{
Generates a plot of the top n words in all the documents in ranges of years. For use with JSTOR's Data for Research datasets (http://dfr.jstor.org/). For best results, repeat the function several times after adding common words to the stopword list and excluding them by re-running the JSTOR_dtmofnouns function. The location of the English stopwords list can be found by entering this at the R prompt: paste0(.libPaths()[1], "/tm/stopwords/english.dat")
}
\examples{
## multiple_archives <- JSTOR_unpack_multiple_archives(mydir = "~/my_data")
}

