% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{JSTORr}
\alias{JSTORr}
\alias{JSTORr-package}
\title{Simple text mining of JSTOR journals}
\description{
Simple text mining and document clustering of JSTOR journal articles
}
\details{
\tabular{ll}{
Package: \tab JSTORr\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-04-25\cr
License: \tab MIT\cr
}

Simple exploratory text mining and document clustering of journal articles from JSTOR's Data for Research service. Go to http://dfr.jstor.org/, make a request for data (specify CSV as outout format and Word Counts as data type), then once you get a zip file, start with the function JSTOR_unpack() and then you're ready to go with any of the other functions. For more details, see https://github.com/UW-ARCHY-textual-macroanalysis-lab/JSTORr
}
\author{
Ben Marwick \email{benmarwick@gmail.com}
}
\references{
\url{http://dfr.jstor.org/}
}

