% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unpack1grams}
\alias{unpack1grams}
\title{unpack1grams}
\format{Returns a list of two items. First is "wordcounts", a Document Term Matrix of 1-grams for 232 articles, and second is 'bibliodata', a data frame of bibliographic information for all 232 articles.}
\source{
\url{http://dfr.jstor.org/}

\url{http://dfr.jstor.org/}
}
\usage{
data(unpack1grams)



unpack2grams

Full-length articles from the Journal of Archaeological Reseach

A dataset containing the full text of full-length articles published in 
the Journal of Archaeological Reseach. Obtained from dfr.jstor.org
}
\description{
Full-length articles from the Journal of Archaeological Reseach
}
\details{
A dataset containing the full text of full-length articles published in 
the Journal of Archaeological Reseach. Obtained from dfr.jstor.org
}

