% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unpack2grams}
\alias{unpack2grams}
\title{unpack2grams}
\format{Returns a list of two items. First is "bigrams", a Document Term Matrix of 2-grams for  articles, and second is 'bibliodata', a data frame of bibliographic information for all articles.}
\source{
\url{http://dfr.jstor.org/}
}
\usage{
data(unpack2grams)
}
\description{
Full-length articles from the Journal of Archaeological Reseach
}
\details{
A dataset containing the full text of full-length articles published in 
the Journal of Archaeological Reseach. Obtained from  \url{http://dfr.jstor.org/}
}

