% Generated by roxygen2 (4.0.2): do not edit by hand
\name{JSTOR_2bigramscor}
\alias{JSTOR_2bigramscor}
\title{Plot the change over time of the correlation between one bigram (or set of bigrams) and another bigram (or set of bigrams) in a JSTOR DfR dataset}
\usage{
JSTOR_2bigramscor(unpack2, bigram1, bigram2, span = 0.4)
}
\arguments{
\item{unpack2}{object returned by the function JSTOR_unpack2.}

\item{bigram1}{One bigram or a vector of bigrams, each bigram surrounded by standard quote marks.}

\item{bigram2}{One bigram or a vector of bigrams, each bigram surrounded by standard quote marks.}

\item{span}{span of the loess line (controls the degree of smoothing). Default is 0.4}
}
\value{
Returns a ggplot object with publication year on the horizontal axis and Pearson's correlation on the vertical axis. Each point represents all the documents of a single year, point size is inversely proportional to p-value of the correlation.
}
\description{
Function to plot changes in the correlation of two sets of bigrams (two sets of 2-grams, or two sets of multiple bigrams) over time. For use with JSTOR's Data for Research datasets (http://dfr.jstor.org/).
}
\examples{
## JSTOR_2bigramscor(unpack2, bigram1 = "hot water", bigram2 = "cold water")
## JSTOR_2bigramscor(unpack2, c("hot water", "warm water", "tepid water"), c("cold water", "ice water"))
}

