% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JSTOR_subset2grams.R
\name{JSTOR_subset2grams}
\alias{JSTOR_subset2grams}
\title{Subset 2gram object}
\usage{
JSTOR_subset2grams(unpack2grams, bigram)
}
\arguments{
\item{bigram}{A bigram, surrounded by standard quote marks.}

\item{unpack1grams}{object returned by the function JSTOR_unpack1grams.}
}
\description{
Subsets bigrams
}

