% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JSTOR_unpack_multiple_archives.R
\name{JSTOR_unpack_multiple_archives}
\alias{JSTOR_unpack_multiple_archives}
\title{Unpacks multiple zip files of JSTOR journal articles and bibliographic data to a Document Term Matrix of 1-grams}
\usage{
JSTOR_unpack_multiple_archives(mydir = getwd())
}
\arguments{
\item{mydir}{path to directory containing multiple zip files dowloaded from dfr.jstor.org (default is the current working directory)}
}
\value{
Returns a list of two items. First is "wordcounts", a Document Term Matrix of 1-grams, and second is 'bibliodata', a data frame of bibliographic information for all articles.
}
\description{
Import journal articles and bibliographic data from multiple downloaded zipfiles, and reshape ready for simple text mining. For use with JSTOR's Data for Research datasets (http://dfr.jstor.org/).
}
\examples{
## multiple_archives <- JSTOR_unpack_multiple_archives(mydir = "~/my_data")
}

